+++ src/Models/X509/BankX509Generator.php
@@ -2,7 +2,6 @@
 
 namespace AndrewSvirin\Ebics\Models\X509;
 
-use AndrewSvirin\Ebics\Contracts\EbicsClientInterface;
 use AndrewSvirin\Ebics\Models\Bank;
 use LogicException;